Reorganize the app into a package
The package directory is 'restaurant_menu_app'. Within the package
direcory, imports and configurations have been moved to '__init__.py'.
Views are now in 'views.py'. The 'templates' and 'static' directories
have also been moved to the package directory. The database is not
included in the repository, but it has also been moved to the package.
Outside the package, 'run.py' contains the 'if __name__ == '__main__'
conditional.