Enable form deletion
The 'forms.py' file now contains a 'DeleteConfirmForm', which features a
radio field with 'Delete' and 'Cancel' options, plus a submit button.
This new form is displayed on the new 'delete_confirm' view. Cancelling
or deleting will send a POST request to the new 'delete' view, which
will either delete the restaurant or not delete the restaurant depending
on the results. The 'delete' view will then redirect the user to the
home page with an appropriate flash message. If the 'delete_confirm'
form does not have an answer selected when the form is submitted, the
post request will be submitted anyway, but the user is redirected back
to the 'delete_confirm' page with a flash message telling the user to
select a radio option.