Enable 'MenuItem' addition
This addition employs 'MenuItemsForm' to receive a menu item name,
course, price, and description from the user. This takes place on the
new 'add_item' view, which returns the new 'add_item.html' template. A
'MenuItemForm' instance is displayed on the template. Restaurant menus
may not have two items by the same name, or a flash message will inform
the user that the user must fill out the form again with a new name.
(This takes place without the 'validate_name' function, which has been
removed.)