Enable 'MenuItem' edits
New 'edit_item' view renders the new 'edit_item.html' template, which
displays a form ('EditItemForm') identical to the 'MenuItemsForm' on the
'add_item.html' template. On a 'GET' request to the view, the form is
pre-filled with existing item information, any of which may be edited.
If the item name is updated, the new name may not match a currently
existing item name for the restaurant, or the user will be redirected to
the same page, and a flash message will tell the user to pick a
different name.