Enable 'MenuItem' deletion
Works exactly like 'Restaurant' deletion. Uses two new views:
'delete_item_confirm' and 'delete_item'. The first of these displays the
'DeleteConfirmForm' with information specific to the item to be deleted.
Once submitted, a POST request is sent to 'delete_item', which will
either delete or not delete, or send the user back to the form page with
to select an option (if the user submits an empty form). In any of these
cases, an appropriate flash message accompanies the form submission.