Enable editing restaurant name
New view, 'edit', allows users to edit the name of a restaurant. The
view displays the 'RestaurantForm', which, on a GET request, is
pre-filled with the current restaurant name. If the name changes, the
new name may not match the name of an existing restaurant. Once
submitted, the database is updated with the new restaurant name.