Update templates
Templates now acknowledge whether a user is logged in. If a user is
logged in, the user will see links for creating a new restaurant or
logging out, and the user will see the user's name at the top of the
home page. If the user is not logged in, a log in link is displayed at
the top of the page.