Fix bug that prevented users from 2nd login
Formerly, users would only be able to log in the first time, and after
logging out, they would be unable to log back in. This was due to a
query error that required the login access token to match the token
stored in the OAuth instance in the database. Following the Flask-Dance
documentation, which had a correct representation of the query, the
token is no longer a filtering requirement.